src/models/model_1/requirements.txt
snowflake-connector-python
python-dotenv
Flask
prometheus_client
